Add bindings for python
The added bindings depend on swig and hence are disabled by
default. If you know what you are doing, enable the bindings
for python. To test the bindings, ensure _libcgroup.so
is copied in the same directory as libcgroup.py. There after
libcgroup API's can be invoked from python.

Please use --enable-bindings after installing swig to compile
with the changes

Signed-off-by: Balbir Singh <balbir@linux.vnet.ibm.com>