cgclear: fixed exit code with -e option
cgclear -e is supposed to clear only empty groups. This works fine, but
when a non-empty group is being cleared, cgclear displays an error and
returns nozero exit code.

It should be silent and return zero exit code, because stumbling upon a
nonempty group is perfectly fine and not an error.

Signed-off-by: Jan Safranek <jsafrane@redhat.com>