cgroup_create_cgroup: write error on fail
In the current version cgroup_create_cgroup returns warning if it fails and
can't set some variable. Thus on default configuration there is no
information about the problem. This situation happens if the cgconfigparser
reads cgconfig.conf. The error message where is the problem is very useful
in this case. The patch do the message visible in standard configuration.

Signed-off-by: Ivana Hutarova Varekova <varekova@redhat.com>
Acked-by: Peter Schiffer <pschiffe@redhat.com>