cgconfig.conf: enable sign "@" in cgroup name
This patch enables to have "@" sign in cgroup name in cgconfig.conf as
well.
It is because of systemd, which uses sign "@" to separate the template name
from the instance identifier in its cgroup names. 
For example: systemd-fsck@.service. 
Without this patch, it's not possible to use these names in cgconfig.conf file.


Signed-off-by: Ivana Hutarova Varekova <varekova@redhat.com>
Acked-by: Peter Schiffer <pschiffe@redhat.com>