tests: Add functional test LXC container class
This commit adds a Container() class for managing LXC containers.
This class provides methods to abstract the management of LXC
containers and simplify their usage in the functional test suite.

Example usages:
  # instantiate the class
  mycontainer = Container('TheNameOfMyContainer')

  # create and start the container
  mycontainer.create()
  mycontainer.start()

Container() can raise ValueError on invalid parameters,
ContainerError when a container operation unexpectedly failes,
or RunError when running a command fails.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Dhaval Giani <dhaval.giani@oracle.com>