README: Add README.md
This commit adds README.md.  It contains graphics that show the
status of the continuous integration and code coverage results.
It also contains a link to the original README.  Note that github
doesn't allow the inlining of files in markdown due to security
concerns, so a link is the best we can do.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Dhaval Giani <dhaval.giani@oracle.com>