tools: Create a cgset library for testing purposes
Create libcgset.la so that functions within cgset can be
unit tested.

A functional test would be preferable in this case, but I'm
looking to test a very specific scenario - where there's an
'=' sign in the value of a setting's name/value pair.  To the
best of my knowledge, only the rdma (v1) and io (v2)
controllers have settings that utilize an '=' in such a fashion.
I can't think of an easy way to test rdma across our test fleet,
and our automated systems don't support the v2 io controller
yet.  Thus a unit test for now.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>