github: Delete the tests folder before bootstrap
On self-hosted github machines, the working directory for
libcgroup is reused from one run to the next.  This is fine
for the root of the directory as the proper git clean/reset
commands are issues.  But this is problematic for the git
submodules like the tests/ folder because clean/reset is nott
run there.

Remove the tests/ directory entirely and rely on ./bootstrap.sh
to recreate and populate it.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>