wrapper.c: Store the version in struct cgroup_controller
Store the cgroup version in struct cgroup_controller.  Each
controller can be mounted separately, and thus the cgroup
version is applicable at the controller level.

Note that this commit requires cgroup_init() to be called by
cgconfigparser.  Storing the version in the cgroup_controller
struct requires the cg_mount_table[] to be populated.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
Reviewed-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>