api.c/wrapper/abstraction: Save previous setting name in control_value
When there's an N->1 mapping, the abstraction layer needs
to know which "1" setting to map back to.  Add a field to
struct control_value that saves off the original setting,
so that the abstraction layer can know how to map back from
an N->1 setting.

A good example of this is cpu.max/cpu.cfs_quota_us.  The
abstraction layer will map cpu.cfs_quota_us to cpu.max.
During the return conversion, the abstraction layer needs
to know whether to map cpu.max to cfs_quota_us or
cfs_period_us.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>