api.c: fix the buffer size to read /proc/cgroups
In cgroup_init(), the first line/header of /proc/cgroups helps in
interpreting the values in the remaining of the file.  We are interested
in the rest of the file and a temporary buffer of size FILENAME_MAX is
allocated/deallocated after reading the header. The length of the header
is only 56 characters in length, excluding the newline. This patch
reduces it to LL_MAX a.k.a 100 characters.

LL_MAX is used instead of introducing another new length macro.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>