samples/c: drop the test suites
samples/c contains test cases, those do serve as good examples on how to
use some APIs, but there are also very elaborate test cases, which are
test suites in themselves.  Let's drop these examples and allow users to
look at gunit/ftests for how to contribute to test cases.  This patch
also modifies the Makefile to help users to build the examples, in case
they want to give it a quick try.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>