tests: Ignore cpuset exclusive errors in test 038
Setting the cpuset exclusive can be challenging because other
cgroups could be using the cpu that is slated to become exclusive
to this cgroup.  Add try/except handling to test 038 for when we
can't get the exclusive bit set, so that the test doesn't fail due
to circumstances beyond libcgroup's control.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>