Support simpler cleanup in functions
Add support for a defer() style function, which cleans up
the memory associated with an auto/local variable when it
goes out of scope. The assumption is that this code is
compiled on a compile with support for the cleanup attribute.

Changes in the patch:
1. Add util.h with support for defer
2. Add common_charptr_cleanup() function to be called for
   freeing char pointers (largely strings in our case containing
   paths)
3. Modify src/api.c to use the new framework

TODO:
1. Support other types of defer functions
2. Support fd closure in cleanup
3. Implement these changes for other files

Testing:
Very light, just compiled and ran simple commands.

Signed-off-by: Balbir singh <bsingharora@gmail.com>