ftests: Use a random container name
Now that we are running github actions against both the
libcgroup and libcgroup-tests repos, it's possible that
collisions can occur when both repos are running their
actions.  (Prior to this commit, the default container
name, TestLibcg, was being used for both repos.)

Use bash's $RANDOM command to generate a random container
name for each run.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
Reviewed-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>