tools/lscgroup: add support for default systemd delegation slice/scope
Enhance the lscgroup tool to support default systemd delegation
slice/scope, if set in the cgconfig.conf's systemd setting.  Parse the
configuration file and read the systemd::delegate setting, if
any, and set it as the default systemd slice/scope in
systemd_default_cgroup. Setting it appends the slice/scope
name to the constructed default cgroup mount path.

When the user passes the relative cgroup name, its appends to the
sub-tree (delegated slice/scope) by default and if the user wishes to
create a cgroup in another subtree or cgroup root hierarchy, they need
to use the absolute path. For example:

                  cgroup_root
                 /          \
                /            \
         systemd.slice*      cgrp1
		|
	 systemd.scope
		|
            cgrp-d1

* default system delegation slice - read from cgconfig.conf
$ sudo lscgroup -g cpu:cgrp-d1

will list the cgroups under the cpu controller hirearchy
cgroup_root:systemd_default_cgroup:cgrp-d1

and for listing a cgroups cgrp1, the user can use:
$ sudo lscgroup -g cpu:/cgrp1

or use the newly introduced -b switch to ignore the systemd slice/scope:
$ sudo lscgroup -b -g cpu:cgrp1

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>