tools/cgclassify: add support for default systemd delegation slice/scope
Enhance the cgclassify tool to support default systemd delegation
slice/scope, if set in the cgconfig.conf's systemd setting.  Parse the
configuration file and read the systemd::delegate setting, if
any, and set it as the default systemd slice/scope in
systemd_default_cgroup. Setting it appends the slice/scope
name to the constructed default cgroup mount path.

When the user passes the relative cgroup name, its appends to the
sub-tree (delegated slice/scope) by default and if the user wishes to
create a cgroup in another subtree or cgroup root hierarchy, they need
to use the absolute path. For example:

                      cgroup_root
                     /          \
                    /            \
             systemd.slice*      cgrp1
                    |
             systemd.scope
                    |
                cgrp-d1

* default system delegation slice/scope - read from cgconfig.conf
$ sudo cgclassify -g cpu:cgrp-d1 1379 (pid)

will create the cgroup under the cpu controller hirearchy
cgroup_root:systemd_default_cgroup:cgrp-d1

and for creating a cgroup cgrp1, the user can use:
$ sudo cgclassify -g cpu:/cgrp1 1379 (pid)

or use the newly introduced -b switch to ignore the systemd slice/scope:
$ sudo cgclassify -b -g cpu:cgrp1 1379 (pid)

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>